week03-local-rag/requirements.txt
gradio
faiss-cpu
dashscope
openai
llama-index-vector-stores-faiss
llama-index-embeddings-dashscope
llama-index-readers-file
docx2txt
openpyxl
llama-index-core
uvicorn
fastapi
llama-index-postprocessor-dashscope-rerank-custom
simplejson
# modelscope==1.18.0
# langchain_community==0.2.16
# transformers==4.44.2
# llama_index.embeddings.huggingface==0.2.3
# llama-index-embeddings-langchain==0.1.2
# datasets==2.21.0
# oss2==2.19.0
# sortedcontainers==2.4.0
# addict==2.4.0s